# perf(compiler-cli): fix regression in type-checking program reuse