--- a/packages/compiler-cli/src/ngtsc/shims/test/reference_tagger_spec.ts
+++ b/packages/compiler-cli/src/ngtsc/shims/test/reference_tagger_spec.ts
@@ -12,6 +12,7 @@ import {absoluteFrom as _, AbsoluteFsPath, getSourceFileOrError} from '../../fil
 import {runInEachFileSystem} from '../../file_system/testing';
 import {makeProgram} from '../../testing';
 import {ShimAdapter} from '../src/adapter';
+import {retagTsFile, untagTsFile} from '../src/expando';
 import {ShimReferenceTagger} from '../src/reference_tagger';
 
 import {TestShimGenerator} from './util';
@@ -67,40 +68,6 @@ runInEachFileSystem(() => {
       expect(shimSf.referencedFiles).toEqual([]);
     });
 
-    it('should remove tags during finalization', () => {
-      const tagger = new ShimReferenceTagger(['test1', 'test2']);
-
-      const fileName = _('/file.ts');
-      const sf = makeArbitrarySf(fileName);
-
-      expectReferencedFiles(sf, []);
-
-      tagger.tag(sf);
-      expectReferencedFiles(sf, ['/file.test1.ts', '/file.test2.ts']);
-
-      tagger.finalize();
-      expectReferencedFiles(sf, []);
-    });
-
-    it('should not remove references it did not add during finalization', () => {
-      const tagger = new ShimReferenceTagger(['test1', 'test2']);
-      const fileName = _('/file.ts');
-      const libFileName = _('/lib.d.ts');
-
-      const sf = makeSf(fileName, `
-        /// <reference path="/lib.d.ts" />
-        export const UNIMPORTANT = true;
-      `);
-
-      expectReferencedFiles(sf, [libFileName]);
-
-      tagger.tag(sf);
-      expectReferencedFiles(sf, ['/file.test1.ts', '/file.test2.ts', libFileName]);
-
-      tagger.finalize();
-      expectReferencedFiles(sf, [libFileName]);
-    });
-
     it('should not tag shims after finalization', () => {
       const tagger = new ShimReferenceTagger(['test1', 'test2']);
       tagger.finalize();
@@ -110,6 +77,56 @@ runInEachFileSystem(() => {
 
       tagger.tag(sf);
       expectReferencedFiles(sf, []);
+    });
+
+    it('should not overwrite original referencedFiles', () => {
+      const tagger = new ShimReferenceTagger(['test']);
+
+      const fileName = _('/file.ts');
+      const sf = makeArbitrarySf(fileName);
+      sf.referencedFiles = [{
+        fileName: _('/other.ts'),
+        pos: 0,
+        end: 0,
+      }];
+
+      tagger.tag(sf);
+      expectReferencedFiles(sf, ['/other.ts', '/file.test.ts']);
+    });
+
+    it('should always tag against the original referencedFiles', () => {
+      const tagger1 = new ShimReferenceTagger(['test1']);
+      const tagger2 = new ShimReferenceTagger(['test2']);
+
+      const fileName = _('/file.ts');
+      const sf = makeArbitrarySf(fileName);
+
+      tagger1.tag(sf);
+      tagger2.tag(sf);
+      expectReferencedFiles(sf, ['/file.test2.ts']);
+    });
+
+    describe('tagging and untagging', () => {
+      it('should be able to untag references and retag them later', () => {
+        const tagger = new ShimReferenceTagger(['test']);
+
+        const fileName = _('/file.ts');
+        const sf = makeArbitrarySf(fileName);
+        sf.referencedFiles = [{
+          fileName: _('/other.ts'),
+          pos: 0,
+          end: 0,
+        }];
+
+        tagger.tag(sf);
+        expectReferencedFiles(sf, ['/other.ts', '/file.test.ts']);
+
+        untagTsFile(sf);
+        expectReferencedFiles(sf, ['/other.ts']);
+
+        retagTsFile(sf);
+        expectReferencedFiles(sf, ['/other.ts', '/file.test.ts']);
+      });
     });
   });
 });