--- a/packages/compiler-cli/src/ngtsc/testing/index.ts
+++ b/packages/compiler-cli/src/ngtsc/testing/index.ts
@@ -5,4 +5,4 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-export {getDeclaration, makeProgram} from './src/utils';
+export {expectCompleteReuse, getDeclaration, makeProgram} from './src/utils';