--- a/packages/compiler-cli/src/ngtsc/typecheck/src/augmented_program.ts
+++ b/packages/compiler-cli/src/ngtsc/typecheck/src/augmented_program.ts
@@ -9,6 +9,7 @@
 import * as ts from 'typescript';
 
 import {AbsoluteFsPath} from '../../file_system';
+import {retagAllTsFiles, untagAllTsFiles} from '../../shims';
 
 import {TypeCheckingProgramStrategy, UpdateMode} from './api';
 import {TypeCheckProgramHost} from './host';
@@ -26,8 +27,10 @@ export class ReusedProgramStrategy implements TypeCheckingProgramStrategy {
    */
   private sfMap = new Map<string, ts.SourceFile>();
 
+  private program: ts.Program = this.originalProgram;
+
   constructor(
-      private program: ts.Program, private originalHost: ts.CompilerHost,
+      private originalProgram: ts.Program, private originalHost: ts.CompilerHost,
       private options: ts.CompilerOptions, private shimExtensionPrefixes: string[]) {}
 
   getProgram(): ts.Program {
@@ -35,6 +38,17 @@ export class ReusedProgramStrategy implements TypeCheckingProgramStrategy {
   }
 
   updateFiles(contents: Map<AbsoluteFsPath, string>, updateMode: UpdateMode): void {
+    if (contents.size === 0) {
+      // No changes have been requested. Is it safe to skip updating entirely?
+      // If UpdateMode is Incremental, then yes. If UpdateMode is Complete, then it's safe to skip
+      // only if there are no active changes already (that would be cleared by the update).
+
+      if (updateMode !== UpdateMode.Complete || this.sfMap.size === 0) {
+        // No changes would be made to the `ts.Program` anyway, so it's safe to do nothing here.
+        return;
+      }
+    }
+
     if (updateMode === UpdateMode.Complete) {
       this.sfMap.clear();
     }
@@ -43,14 +57,25 @@ export class ReusedProgramStrategy implements TypeCheckingProgramStrategy {
       this.sfMap.set(filePath, ts.createSourceFile(filePath, text, ts.ScriptTarget.Latest, true));
     }
 
-    const host =
-        new TypeCheckProgramHost(this.sfMap, this.originalHost, this.shimExtensionPrefixes);
+    const host = new TypeCheckProgramHost(
+        this.sfMap, this.originalProgram, this.originalHost, this.shimExtensionPrefixes);
+    const oldProgram = this.program;
+
+    // Retag the old program's `ts.SourceFile`s with shim tags, to allow TypeScript to reuse the
+    // most data.
+    retagAllTsFiles(oldProgram);
+
     this.program = ts.createProgram({
       host,
       rootNames: this.program.getRootFileNames(),
       options: this.options,
-      oldProgram: this.program,
+      oldProgram,
     });
     host.postProgramCreationCleanup();
+
+    // And untag them afterwards. We explicitly untag both programs here, because the oldProgram
+    // may still be used for emit and needs to not contain tags.
+    untagAllTsFiles(this.program);
+    untagAllTsFiles(oldProgram);
   }
 }