… packages/compiler-cli/test/ngtsc/env.ts
… packages/compiler-cli/test/ngtsc/template_typecheck_spec.ts