@alan-agius4 2020-06-22
@kyliau, requested to gather some rebuild timings for this change.

### Important notes
- The below timings have been gather on AIO, using the latest RC’s for both Angular FW and CLI.  
- The first timing is that of the first incremental re-build and not of the cold build.
- The first rebuild with fork type checker is typically slower than the following because a program nor source files have been created and cached yet in the forked process.

### With fork type checker

```
ng build --configuration=fast --aot --watch --progress --fork-type-checker
```

**Current**
```
[benchmark]   Average Time: 5014.60 ms (18230.00, 1709.00, 1409.00, 1511.00, 2214.00)
[benchmark]   Average Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Peak Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Average CPU usage: 192.48 % (215.55, 239.11, 183.33, 179.83, 144.59)
[benchmark]   Peak CPU usage: 398.44 % (610.00, 422.22, 360.00, 340.00, 260.00)
[benchmark]   Average Memory usage: 1338.61 MB (959.83, 1373.54, 1428.99, 1435.42, 1495.27)
[benchmark]   Peak Memory usage: 1489.55 MB (1556.14, 1418.00, 1456.09, 1479.27, 1538.27)
```

**With this change**
```
[benchmark]   Averaqe Time: 4894.00 ms (18734.00, 1510.00, 1410.00, 1309.00, 1507.00)
[benchmark]   Average Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Peak Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Average CPU usage: 184.63 % (208.48, 210.00, 178.07, 172.67, 153.91)
[benchmark]   Peak CPU usage: 420.00 % (640.00, 370.00, 370.00, 360.00, 360.00)
[benchmark]   Average Memory usage: 1118.72 MB (991.59, 1100.26, 1153.92, 1149.76, 1198.05)
[benchmark]   Peak Memory usage: 1278.98 MB (1658.93, 1145.76, 1160.33, 1182.18, 1247.69)
```



### Without fork type checker

```
ng build --configuration=fast --aot --watch --progress --no-fork-type-checker
```

**Current**
```
[benchmark]   Average Time: 1952.20 ms (2282.00, 2077.00, 1775.00, 1859.00, 1768.00)
[benchmark]   Average Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Peak Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Average CPU usage: 192.48 % (215.55, 239.11, 183.33, 179.83, 144.59)
[benchmark]   Peak CPU usage: 398.44 % (610.00, 422.22, 360.00, 340.00, 260.00)
[benchmark]   Average Memory usage: 1338.61 MB (959.83, 1373.54, 1428.99, 1435.42, 1495.27)
[benchmark]   Peak Memory usage: 1489.55 MB (1556.14, 1418.00, 1456.09, 1479.27, 1538.27)
```

**With this change**
```
[benchmark]   Averaqe Time: 1891.00 ms (2188.00, 1878.00, 1859.00, 1760.00, 1770.00)
[benchmark]   Average Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Peak Process usage: 2.00 process(es) (2.00, 2.00, 2.00, 2.00, 2.00)
[benchmark]   Average CPU usage: 184.63 % (208.48, 210.00, 178.07, 172.67, 153.91)
[benchmark]   Peak CPU usage: 420.00 % (640.00, 370.00, 370.00, 360.00, 360.00)
[benchmark]   Average Memory usage: 1118.72 MB (991.59, 1100.26, 1153.92, 1149.76, 1198.05)
[benchmark]   Peak Memory usage: 1278.98 MB (1658.93, 1145.76, 1160.33, 1182.18, 1247.69)
```