@kyliau 2020-06-22
Thank you @alan-agius4. Looking at the numbers, there is indeed an improvement in both time and memory usage, but they are pretty small. This gives us confidence that there is not significant regression in the external toolchain, therefore this commit is not v10.0.0 blocking. This will likely be released in v10.0.1.